Administrador/backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
psycopg2-binary==2.9.9
python-decouple==3.8
pillow==10.1.0
django-filter==23.4
tensorflow==2.14.0
keras==2.14.0
numpy==1.24.3
pandas==2.1.3
python-dotenv==1.0.0
requests==2.31.0
Werkzeug==3.0.1